#110 move pipline to the separate file and added totalOffer
